Refactor ingredient displays and shared logic
- Replace inline ingredient image logic in `CraftedIngredientDisplay.tsx`.
- Simplify `FarmedIngredientDisplay.tsx` with shared image hook.
- Create `useIngredientImage.ts` hook for ingredient image logic.
- Add `formatIngredientName` utility in `src/utils/utils.ts`.
